Separated Application and Client API into their own classes, added GetServerById to Application API
