+++ Sharpdactyl/Sharpdactyl.csproj
@@ -8,7 +8,7 @@
     <PackageTags>Library</PackageTags>
     <Description>API library for the Pterodactyl game panel</Description>
     <PackageTags>API;Library; Pterodactyl Game Panel; Ptero;</PackageTags>
-    <PackageReleaseNotes>Version 0.0.2-alpha</PackageReleaseNotes>
+    <PackageReleaseNotes>Version 1.0.2</PackageReleaseNotes>
     <PackageProjectUrl>https://github.com/JMWren/Sharpdactyl</PackageProjectUrl>
     <PackageLicenseExpression>MIT</PackageLicenseExpression>
     <PackageRequireLicenseAcceptance>false</PackageRequireLicenseAcceptance>
